align module names to the route definitions
