+++ src/lean_spec/subspecs/api/endpoints/__init__.py
@@ -1,18 +1,10 @@
-"""
-API endpoint specifications.
+"""API endpoint specifications."""
 
-Each module defines the spec (routes, constants, response formats) and handlers
-for a specific endpoint:
-
-- health: Health check endpoint
-- checkpoint_sync: Finalized state and justified checkpoint endpoints
-- metrics: Prometheus metrics endpoint (optional)
-"""
-
-from . import checkpoint_sync, health, metrics
+from . import checkpoints, health, metrics, states
 
 __all__ = [
-    "checkpoint_sync",
+    "checkpoints",
     "health",
     "metrics",
+    "states",
 ]
