+++ src/lean_spec/subspecs/api/endpoints/states.py
@@ -1,14 +1,8 @@
-"""
-Checkpoint sync endpoint specifications and handlers.
-
-Used for checkpoint sync - clients download finalized state to bootstrap quickly
-instead of syncing from genesis.
-"""
+"""States endpoint handlers."""
 
 from __future__ import annotations
 
 import asyncio
-import json
 import logging
 
 from aiohttp import web
@@ -16,7 +10,7 @@ from aiohttp import web
 logger = logging.getLogger(__name__)
 
 
-async def handle_finalized_state(request: web.Request) -> web.Response:
+async def handle_finalized(request: web.Request) -> web.Response:
     """
     Handle finalized state request.
 
@@ -50,36 +44,3 @@ async def handle_finalized_state(request: web.Request) -> web.Response:
         raise web.HTTPInternalServerError(reason="Encoding failed") from e
 
     return web.Response(body=ssz_bytes, content_type="application/octet-stream")
-
-
-async def handle_justified_checkpoint(request: web.Request) -> web.Response:
-    """
-    Handle justified checkpoint request.
-
-    Returns the latest justified checkpoint for monitoring consensus progress.
-
-    Response: JSON object with fields:
-        - slot (integer): The slot number of the justified checkpoint.
-        - root (string): The block root as 0x-prefixed hex string (66 chars total).
-
-    Status Codes:
-        200 OK: Checkpoint returned successfully.
-        503 Service Unavailable: Store not initialized.
-    """
-    store_getter = request.app.get("store_getter")
-    store = store_getter() if store_getter else None
-
-    if store is None:
-        raise web.HTTPServiceUnavailable(reason="Store not initialized")
-
-    justified = store.latest_justified
-
-    return web.Response(
-        body=json.dumps(
-            {
-                "slot": justified.slot,
-                "root": "0x" + justified.root.hex(),
-            }
-        ),
-        content_type="application/json",
-    )
